+++ client/display/mod.rs
@@ -1,12 +1,11 @@
-use comfy_table::Color;
-
+pub mod colors;
 mod follow;
 mod group;
 pub mod helper;
 mod log;
 mod state;
 
-use self::helper::style_text;
+use self::{colors::Colors, helper::style_text};
 
 // Re-exports
 pub use self::follow::follow_local_task_logs;
@@ -15,12 +14,12 @@ pub use self::log::print_logs;
 pub use self::state::print_state;
 
 /// Used to style any generic success message from the daemon.
-pub fn print_success(message: &str) {
+pub fn print_success(_colors: &Colors, message: &str) {
     println!("{}", message);
 }
 
 /// Used to style any generic failure message from the daemon.
-pub fn print_error(message: &str) {
-    let styled = style_text(message, Some(Color::Red), None);
+pub fn print_error(colors: &Colors, message: &str) {
+    let styled = style_text(message, Some(colors.red()), None);
     println!("{}", styled);
 }